Move to quadcopter naming
