+++ src/gams/platforms/airlib/AirLibQuadcopterBase.h
